Inject lazy dependencies
Still TODO Find a nice way to expose WebampLazy as:

```JavaScript
import WebampLazy from 'webamp/lazy';
```
