+++ config/webpack.library.js
@@ -7,7 +7,8 @@ module.exports = {
   node: {
     // Consider suggesting jsmediatags use: https://github.com/feross/is-buffer
     // Cuts 22k
-    Buffer: false
+    Buffer: false,
+    fs: "empty"
   },
   module: {
     rules: [
@@ -64,7 +65,9 @@ module.exports = {
   ],
   entry: {
     bundle: "./js/webamp.js",
-    "bundle.min": "./js/webamp.js"
+    "bundle.min": "./js/webamp.js",
+    "lazy-bundle": "./js/webampLazy.js",
+    "lazy-bundle.min": "./js/webampLazy.js"
   },
   output: {
     path: path.resolve(__dirname, "../built"),
