+++ package.json
@@ -6,7 +6,8 @@
   "files": [
     "built/webamp.bundle.js",
     "built/webamp.bundle.min.js",
-    "index.d.ts"
+    "built/webamp-lazy.bundle.js",
+    "built/webamp-lazy.bundle.min.js"
   ],
   "scripts": {
     "lint": "eslint .",
